:sparkles: separate tab for map in mobile screen
