+++ src/pages/detail-page/features/SummaryAndDownloadPanel.tsx
@@ -1,83 +1,27 @@
-import React, {
-  FC,
-  startTransition,
-  useCallback,
-  useMemo,
-  useState,
-} from "react";
+import React, { FC, useMemo } from "react";
 import { Box, Grid, Stack } from "@mui/material";
-import { padding } from "../../../styles/constants";
 import { useDetailPageContext } from "../context/detail-page-context";
-import Controls from "../../../components/map/mapbox/controls/Controls";
-import NavigationControl from "../../../components/map/mapbox/controls/NavigationControl";
-import ScaleControl from "../../../components/map/mapbox/controls/ScaleControl";
-import MapBox from "../../../components/map/mapbox/Map";
-import Layers from "../../../components/map/mapbox/layers/Layers";
 import ExpandableTextArea from "../../../components/list/listItem/subitem/ExpandableTextArea";
-import DrawRect from "../../../components/map/mapbox/controls/menu/DrawRect";
-import { LngLatBounds, MapEvent } from "mapbox-gl";
-import BaseMapSwitcher from "../../../components/map/mapbox/controls/menu/BaseMapSwitcher";
-import MenuControl from "../../../components/map/mapbox/controls/menu/MenuControl";
-import DateRange from "../../../components/map/mapbox/controls/menu/DateRange";
 import dayjs, { Dayjs } from "dayjs";
-import {
-  DateRangeCondition,
-  DownloadConditionType,
-  DownloadServiceType,
-  SubsettingType,
-  BBoxCondition,
-  PolygonCondition,
-} from "../context/DownloadDefinitions";
-import { dateDefault } from "../../../components/common/constants";
-import {
-  FeatureCollection,
-  Point,
-  Feature,
-  Polygon,
-  MultiPolygon,
-} from "geojson";
-import DisplayCoordinate from "../../../components/map/mapbox/controls/DisplayCoordinate";
-import HexbinLayer from "../../../components/map/mapbox/layers/HexbinLayer";
-import GeoServerLayer, {
-  Dimension,
-} from "../../../components/map/mapbox/layers/GeoServerLayer";
-import MapLayerSwitcher, {
-  LayerName,
-  LayerSwitcherLayer,
-} from "../../../components/map/mapbox/controls/menu/MapLayerSwitcher";
+import { FeatureCollection, Point } from "geojson";
 import {
   DatasetType,
   OGCCollection,
 } from "../../../components/common/store/OGCCollectionDefinitions";
-import ReferenceLayerSwitcher, {
-  staticBaseLayerConfig,
-} from "../../../components/map/mapbox/controls/menu/ReferenceLayerSwitcher";
-import MenuControlGroup from "../../../components/map/mapbox/controls/menu/MenuControlGroup";
-import GeojsonLayer from "../../../components/map/mapbox/layers/GeojsonLayer";
+import {
+  LayerName,
+  LayerSwitcherLayer,
+} from "../../../components/map/mapbox/controls/menu/MapLayerSwitcher";
 import useBreakpoint from "../../../hooks/useBreakpoint";
-import FitToSpatialExtentsLayer from "../../../components/map/mapbox/layers/FitToSpatialExtentsLayer";
 import AIGenTag from "../../../components/info/AIGenTag";
-import { MapEventEnum } from "../../../components/map/mapbox/constants";
-import { DateSliderPoint } from "../../../components/common/slider/DateSlider";
-import { dateToValue } from "../../../utils/DateUtils";
 import { portalTheme } from "../../../styles";
-import { GeoserverFieldsResponse } from "../../../components/common/store/GeoserverDefinitions";
-import * as turf from "@turf/turf";
-import _ from "lodash";
-import { createStaticLayers } from "../../../components/map/mapbox/layers/StaticLayer";
 
-const mapContainerId = "map-detail-container-id";
-
-interface SummaryAndDownloadPanelProps {
-  mapFocusArea?: LngLatBounds;
-  onMapMoveEnd?: (evt: MapEvent) => void;
-}
-
-const getMinMaxDateStamps = (
+// Exported for unit tests
+export const getMinMaxDateStamps = (
   featureCollection?: FeatureCollection<Point>
 ): [Dayjs, Dayjs] => {
   if (!featureCollection?.features?.length) {
-    return [dayjs(dateDefault.min), dayjs(dateDefault.max)];
+    return [dayjs(new Date("01/01/1970")), dayjs(new Date())];
   }
 
   let minDate: Dayjs | null = null;
@@ -99,12 +43,12 @@ const getMinMaxDateStamps = (
   }
 
   return [
-    minDate && minDate.isValid() ? minDate : dayjs(dateDefault.min),
-    maxDate && maxDate.isValid() ? maxDate : dayjs(dateDefault.max),
+    minDate && minDate.isValid() ? minDate : dayjs(new Date("01/01/1970")),
+    maxDate && maxDate.isValid() ? maxDate : dayjs(new Date()),
   ];
 };
 
-// We want to vitest this easier
+// Exported for unit tests and MapPanel
 export const buildMapLayerConfig = (
   collection: OGCCollection | null | undefined,
   hasSummaryFeature: boolean,
@@ -116,18 +60,13 @@ export const buildMapLayerConfig = (
   const layers: LayerSwitcherLayer<LayerName>[] = [];
 
   if (collection) {
-    // Only show hexbin layer when the collection has summary feature and it is NOT a zarr dataset
     const isSupportHexbin = hasSummaryFeature && !isZarrDataset;
 
-    // Show spatial extent layer when
-    // 1. it is a zarr dataset
-    // 2. or no geoserver layer nor hexbin layer
     const isSupportSpatialExtent =
       hasSpatialExtent &&
       ((hasSummaryFeature && isZarrDataset) ||
         (!isWMSAvailable && !isSupportHexbin));
 
-    // Must be ordered by Hexbin > GeoServer > Spatial extents
     if (isSupportHexbin) {
       const l: LayerSwitcherLayer<LayerName> = {
         id: LayerName.Hexbin,
@@ -141,7 +80,6 @@ export const buildMapLayerConfig = (
       const l: LayerSwitcherLayer<LayerName> = {
         id: LayerName.GeoServer,
         name: "Geoserver",
-        // If hexbin is supported, then geoserver is not default
         selected: !isSupportHexbin,
       };
       layers.push(l);
@@ -157,324 +95,24 @@ export const buildMapLayerConfig = (
     }
 
     if (lastSelectedLayer) {
-      // Check if we have last selected layer, if yes, select it, otherwise select the first one
       layers.forEach((l) => (l.selected = l.id === lastSelectedLayer.id));
     } else if (
       layers.length > 0 &&
       layers.find((l) => l.selected) === undefined
     ) {
-      // If non of the layer is set to selected, select the first one
       layers[0].selected = true;
     }
   }
   return layers;
 };
 
-const SummaryAndDownloadPanel: FC<SummaryAndDownloadPanelProps> = ({
-  mapFocusArea,
-  onMapMoveEnd,
-}) => {
-  const {
-    collection,
-    featureCollection,
-    downloadConditions,
-    getAndSetDownloadConditions,
-    lastSelectedMapLayer,
-    setLastSelectedMapLayer,
-    selectedWmsLayer,
-    setSelectedWmsLayer,
-    downloadService,
-    selectedCoKey,
-    setSelectedCoKey,
-  } = useDetailPageContext();
+const SummaryAndDownloadPanel: FC = () => {
+  const { collection } = useDetailPageContext();
+  const { isUnderLaptop, isMobile } = useBreakpoint();
 
-  // Need to init with null as collection value can be undefined when it entered this component.
-  const [mapLayerConfig, setMapLayerConfig] = useState<
-    LayerSwitcherLayer<LayerName>[]
-  >([]);
-  const [staticLayer, setStaticLayer] = useState<Array<string>>([]);
-  const [isWMSAvailable, setIsWMSAvailable] = useState<boolean>(true);
-  const [_wmsFields, setWMSFields] = useState<GeoserverFieldsResponse[]>([]);
-  const [timeSliderSupport, setTimeSliderSupport] = useState<boolean>(false);
-  const [drawRectSupport, setDrawRectSupportSupport] = useState<boolean>(false);
-  const [discreteTimeSliderValues, setDiscreteTimeSliderValues] = useState<
-    Map<string, Array<number>> | undefined
-  >(undefined);
-  const [datePointValue, setDatePointValue] = useState<number>(
-    dateToValue(dayjs(dateDefault.min))
-  );
-  const { isUnderLaptop } = useBreakpoint();
-
-  const [
-    abstract,
-    hasSummaryFeature,
-    noMapPreview,
-    minDateStamp,
-    maxDateStamp,
-  ] = useMemo(() => {
-    const abstract =
-      collection?.getEnhancedDescription() || collection?.description || "";
-    const hasSummaryFeature = collection?.hasSummaryFeature() || false;
-    const bbox = collection?.getBBox();
-    const hasSpatialExtent = Array.isArray(bbox) && bbox.length > 0;
-    const isZarrDataset = collection?.getDatasetType() === DatasetType.ZARR;
-
-    const scope = collection?.getScope();
-    const isDocumentScope = scope?.toLowerCase() === "document";
-    // for document records, no map preview if there is no spatial extent,
-    // for dataset records, no map preview if satisfies both: no co download and no wfs download service and no spatial extent
-    const noMapPreview = isDocumentScope
-      ? !hasSpatialExtent
-      : downloadService === DownloadServiceType.Unavailable &&
-        !hasSpatialExtent &&
-        !isWMSAvailable;
-    // We trust the metadata value instead of raw data, in fact it is hard to have a common
-    // time value, for example cloud optimized date range may be different from the
-    // geoserver one
-    let start = dayjs(dateDefault.min);
-    let end = dayjs(dateDefault.max);
-
-    const extent = collection?.getExtent();
-    if (extent) {
-      const [s, e] = extent.getOverallTemporal();
-      start = s === undefined ? start : dayjs(s, dateDefault.DISPLAY_FORMAT);
-      end = e === undefined ? end : dayjs(e, dateDefault.DISPLAY_FORMAT);
-    }
-
-    startTransition(() => {
-      setMapLayerConfig(
-        buildMapLayerConfig(
-          collection,
-          hasSummaryFeature,
-          isZarrDataset,
-          isWMSAvailable,
-          hasSpatialExtent,
-          lastSelectedMapLayer
-        )
-      );
-    });
-
-    return [abstract, hasSummaryFeature, noMapPreview, start, end];
-  }, [collection, downloadService, isWMSAvailable, lastSelectedMapLayer]);
-
-  const [filterStartDate, filterEndDate] = useMemo(() => {
-    const dateRangeConditionGeneric = downloadConditions.find(
-      (condition) => condition.type === DownloadConditionType.DATE_RANGE
-    );
-    if (!dateRangeConditionGeneric) {
-      return [undefined, undefined];
-    }
-    const dateRangeCondition = dateRangeConditionGeneric as DateRangeCondition;
-
-    const conditionStart = dayjs(
-      dateRangeCondition.start,
-      dateDefault.DATE_FORMAT
-    );
-    const conditionEnd = dayjs(dateRangeCondition.end, dateDefault.DATE_FORMAT);
-    return [conditionStart, conditionEnd];
-  }, [downloadConditions]);
-
-  const drawFeatures = useMemo(() => {
-    const existingBboxConditions = downloadConditions.filter(
-      (condition) => condition.type === DownloadConditionType.BBOX
-    ) as BBoxCondition[];
-
-    const existingPolygonConditions = downloadConditions.filter(
-      (condition) => condition.type === DownloadConditionType.POLYGON
-    ) as PolygonCondition[];
-
-    const features: Feature<Polygon>[] = [];
-
-    existingBboxConditions.forEach((condition) => {
-      const [west, south, east, north] = condition.bbox;
-      features.push({
-        type: "Feature",
-        geometry: {
-          type: "Polygon",
-          coordinates: [
-            [
-              [west, north],
-              [east, north],
-              [east, south],
-              [west, south],
-              [west, north],
-            ],
-          ],
-        },
-        properties: { selectionType: "bbox" },
-      });
-    });
-
-    existingPolygonConditions.forEach((condition) => {
-      const closedCoords = [...condition.coordinates, condition.coordinates[0]];
-      features.push({
-        type: "Feature",
-        geometry: {
-          type: "Polygon",
-          coordinates: [closedCoords],
-        },
-        properties: { selectionType: "polygon" },
-      });
-    });
-
-    return features;
-  }, [downloadConditions]);
-
-  const geoServerLayerConfig = useMemo(() => {
-    return discreteTimeSliderValues
-      ? {
-          urlParams: {
-            TIME: dayjs.utc(datePointValue!),
-            MODE: Dimension.SINGLE,
-          },
-        }
-      : {
-          urlParams: {
-            START_DATE: filterStartDate,
-            END_DATE: filterEndDate,
-          },
-        };
-  }, [
-    discreteTimeSliderValues,
-    datePointValue,
-    filterStartDate,
-    filterEndDate,
-  ]);
-
-  const handleMapChange = useCallback(
-    (event: MapEvent | undefined) => {
-      // implement later
-      event && event.type === MapEventEnum.MOVE_END && onMapMoveEnd?.(event);
-    },
-    [onMapMoveEnd]
-  );
-
-  const checkSubsettingSupport = useCallback(
-    (subsettingType: SubsettingType) => {
-      switch (subsettingType) {
-        // Time slider support when
-        // 1. CO download is available
-        // 2. or the in geoserver and it supports time slider
-        case SubsettingType.TimeSlider:
-          return (
-            hasSummaryFeature ||
-            (mapLayerConfig.filter((m) => m.selected)?.[0]?.id ===
-              LayerName.GeoServer &&
-              timeSliderSupport)
-          );
-
-        // Draw rect support when
-        // 1. CO download is available
-        // 2. or the in geoserver and it supports draw rect
-        // Also the download service must be available
-        case SubsettingType.DrawRect:
-          return (
-            (hasSummaryFeature ||
-              (mapLayerConfig.filter((m) => m.selected)?.[0]?.id ===
-                LayerName.GeoServer &&
-                drawRectSupport)) &&
-            downloadService !== DownloadServiceType.Unavailable
-          );
-        default:
-          return false;
-      }
-    },
-    [
-      hasSummaryFeature,
-      mapLayerConfig,
-      timeSliderSupport,
-      drawRectSupport,
-      downloadService,
-    ]
-  );
-
-  const handleMapLayerChange = useCallback(
-    (layer: LayerSwitcherLayer<LayerName>) => {
-      layer.selected = true;
-      setLastSelectedMapLayer(layer);
-      setMapLayerConfig((prevLayers: LayerSwitcherLayer<LayerName>[]) => {
-        const layers = [...prevLayers];
-        layers.forEach((l) => (l.selected = l.id === layer.id));
-        return layers;
-      });
-    },
-    [setLastSelectedMapLayer]
-  );
-
-  const handleSliderPointChange = useCallback(
-    (
-      event: Event | React.SyntheticEvent<Element, Event> | undefined,
-      value: number | number[]
-    ) => {
-      setDatePointValue(value as number);
-    },
-    []
-  );
-
-  const onWMSAvailabilityChange = useCallback((isWMSAvailable: boolean) => {
-    setIsWMSAvailable(isWMSAvailable);
-  }, []);
-
-  const handleBaseMapSwitch = useCallback(
-    (target: EventTarget & HTMLInputElement) =>
-      setStaticLayer((values) => {
-        // Remove the item and add it back if selected
-        const e = values?.filter((i) => i !== target.value);
-        if (target.checked) {
-          e.push(target.value);
-        }
-        return [...e];
-      }),
-    []
-  );
-
-  const onWmsLayerChange = useCallback(
-    (wmsLayerName: string) => setSelectedWmsLayer(wmsLayerName),
-    [setSelectedWmsLayer]
-  );
-
-  const handleFeaturesChange = useCallback(
-    (
-      newFeatures: Feature<Polygon | MultiPolygon>[],
-      removeFeature: (id: string) => void
-    ) => {
-      const bboxConditions: BBoxCondition[] = [];
-      const polygonConditions: PolygonCondition[] = [];
-
-      newFeatures.forEach((feature) => {
-        const id = String(feature.id);
-        const selectionType = feature.properties?.selectionType || "bbox";
-        const removeCallback = () => removeFeature(id);
-
-        if (selectionType === "polygon") {
-          const coords = feature.geometry.coordinates[0];
-          // Remove closing duplicate point if present
-          const vertices =
-            coords.length > 1 &&
-            coords[0][0] === coords[coords.length - 1][0] &&
-            coords[0][1] === coords[coords.length - 1][1]
-              ? coords.slice(0, -1)
-              : coords;
-          polygonConditions.push(
-            new PolygonCondition(
-              id,
-              vertices as [number, number][],
-              removeCallback
-            )
-          );
-        } else {
-          const bbox = turf.bbox(feature);
-          bboxConditions.push(new BBoxCondition(id, bbox, removeCallback));
-        }
-      });
-
-      getAndSetDownloadConditions(DownloadConditionType.BBOX, bboxConditions);
-      getAndSetDownloadConditions(
-        DownloadConditionType.POLYGON,
-        polygonConditions
-      );
-    },
-    [getAndSetDownloadConditions]
+  const abstract = useMemo(
+    () => collection?.getEnhancedDescription() || collection?.description || "",
+    [collection]
   );
 
   return (
@@ -494,6 +132,7 @@ const SummaryAndDownloadPanel: FC<SummaryAndDownloadPanelProps> = ({
               <ExpandableTextArea
                 text={abstract}
                 showMoreStr={"Show All"}
+                defaultExpanded={isMobile}
                 sx={{
                   width: isUnderLaptop ? "95%" : "98%",
                   "& button": {
@@ -504,137 +143,6 @@ const SummaryAndDownloadPanel: FC<SummaryAndDownloadPanelProps> = ({
                 }}
               />
             </Stack>
-            <Box
-              arial-label="map"
-              id={mapContainerId}
-              sx={{
-                width: "100%",
-                minHeight: "550px",
-                marginY: padding.large,
-              }}
-            >
-              <MapBox
-                animate={false}
-                panelId={mapContainerId}
-                projection={"mercator"} // Hexbin support this project or globe only
-                announcement={
-                  noMapPreview ? "Dataset preview is not available" : undefined
-                }
-                onMoveEvent={handleMapChange}
-                onZoomEvent={handleMapChange}
-              >
-                <Controls>
-                  <NavigationControl visible={!isUnderLaptop} />
-                  <ScaleControl />
-                  <DisplayCoordinate />
-                  <MenuControlGroup>
-                    <MenuControl menu={<BaseMapSwitcher />} />
-                    <MenuControl
-                      menu={
-                        <ReferenceLayerSwitcher
-                          layers={staticBaseLayerConfig}
-                          onEvent={handleBaseMapSwitch}
-                        />
-                      }
-                    />
-                    <MenuControl
-                      menu={
-                        <MapLayerSwitcher
-                          layers={mapLayerConfig}
-                          onEvent={handleMapLayerChange}
-                        />
-                      }
-                      visible={mapLayerConfig.length !== 0}
-                    />
-                    <MenuControl
-                      visible={checkSubsettingSupport(
-                        SubsettingType.TimeSlider
-                      )}
-                      menu={
-                        <DateRange
-                          key="date-range"
-                          minDate={minDateStamp.format(dateDefault.DATE_FORMAT)}
-                          maxDate={maxDateStamp.format(dateDefault.DATE_FORMAT)}
-                          getAndSetDownloadConditions={
-                            getAndSetDownloadConditions
-                          }
-                          downloadConditions={downloadConditions}
-                          options={
-                            discreteTimeSliderValues
-                              ? {
-                                  additionalSlider: (
-                                    <DateSliderPoint
-                                      valid_points={discreteTimeSliderValues?.get(
-                                        selectedWmsLayer
-                                      )}
-                                      onDatePointChange={
-                                        handleSliderPointChange
-                                      }
-                                    />
-                                  ),
-                                }
-                              : undefined
-                          }
-                        />
-                      }
-                    />
-                    <MenuControl
-                      visible={checkSubsettingSupport(SubsettingType.DrawRect)}
-                      menu={
-                        <DrawRect
-                          onChangeFeatures={handleFeaturesChange}
-                          features={drawFeatures}
-                        />
-                      }
-                    />
-                  </MenuControlGroup>
-                </Controls>
-                <Layers>
-                  <FitToSpatialExtentsLayer
-                    collection={collection}
-                    bbox={mapFocusArea}
-                  />
-                  {createStaticLayers(staticLayer)}
-                  {
-                    // Put the first two later here so that they all init the same time
-                    // GeoServerLayer is heavy to load, so we can load it
-                    // but hide it with visible = false
-                  }
-                  <HexbinLayer
-                    featureCollection={featureCollection}
-                    filterStartDate={filterStartDate}
-                    filterEndDate={filterEndDate}
-                    visible={
-                      mapLayerConfig.filter((m) => m?.selected)?.[0]?.id ===
-                      LayerName.Hexbin
-                    }
-                    selectedCoKey={selectedCoKey}
-                    onSelectCoKey={setSelectedCoKey}
-                  />
-                  <GeoServerLayer
-                    geoServerLayerConfig={geoServerLayerConfig}
-                    onWMSAvailabilityChange={onWMSAvailabilityChange}
-                    onWmsLayerChange={onWmsLayerChange}
-                    setWmsFields={setWMSFields}
-                    setTimeSliderSupport={setTimeSliderSupport}
-                    setDiscreteTimeSliderValues={setDiscreteTimeSliderValues}
-                    setDrawRectSupportSupport={setDrawRectSupportSupport}
-                    collection={collection}
-                    visible={
-                      mapLayerConfig.filter((m) => m?.selected)?.[0]?.id ===
-                      LayerName.GeoServer
-                    }
-                  />
-                  <GeojsonLayer
-                    collection={collection}
-                    visible={
-                      mapLayerConfig.filter((m) => m?.selected)?.[0]?.id ===
-                      LayerName.SpatialExtent
-                    }
-                  />
-                </Layers>
-              </MapBox>
-            </Box>
           </Stack>
         </Grid>
       </Grid>
@@ -642,5 +150,4 @@ const SummaryAndDownloadPanel: FC<SummaryAndDownloadPanelProps> = ({
   );
 };
 
-export { getMinMaxDateStamps };
 export default SummaryAndDownloadPanel;
