+++ src/pages/detail-page/layout/ContentSection.tsx
@@ -4,11 +4,13 @@ import AdditionalInfoPanel from "../features/AdditionalInfoPanel";
 import CitationPanel from "../features/CitationPanel";
 import SummaryAndDownloadPanel from "../features/SummaryAndDownloadPanel";
 import AssociatedRecordsPanel from "../features/AssociatedRecordsPanel";
+import MapPanel from "../features/MapPanel";
 import { IconSummary } from "../../../components/icon/tabs/IconSummary";
 import { IconDataAccess } from "../../../components/icon/tabs/IconDataAccess";
 import { IconCitation } from "../../../components/icon/tabs/IconCitation";
 import { IconInformation } from "../../../components/icon/tabs/IconInformation3";
 import { IconRelatedResources } from "../../../components/icon/tabs/IconRelatedResources";
+import { IconMap } from "../../../components/icon/tabs/IconMap";
 import { Box, Card } from "@mui/material";
 import { borderRadius } from "../../../styles/constants";
 import { useDetailPageContext } from "../context/detail-page-context";
@@ -22,6 +24,7 @@ import {
   pageReferer,
 } from "../../../components/common/constants";
 import useTabNavigation from "../../../hooks/useTabNavigation";
+import useBreakpoint from "../../../hooks/useBreakpoint";
 import notMatchingRecordImage from "@/assets/images/no_matching_record.png";
 
 interface ContentSectionProps {
@@ -63,22 +66,39 @@ const associatedRecordsPanelTab: Tab = {
   icon: <IconRelatedResources />,
 };
 
-const summaryAndDownloadPanelTab = (
+const summaryOnlyTab: Tab = {
+  label: "Summary",
+  value: detailPageDefault.SUMMARY,
+  component: <SummaryAndDownloadPanel />,
+  icon: <IconSummary />,
+};
+
+const summaryWithMapTab = (
   mapFocusArea: LngLatBounds | undefined,
   onMapMoveEnd?: (evt: MapEvent) => void
-): Tab => {
-  return {
-    label: "Summary",
-    value: detailPageDefault.SUMMARY,
-    component: (
-      <SummaryAndDownloadPanel
-        mapFocusArea={mapFocusArea}
-        onMapMoveEnd={onMapMoveEnd}
-      />
-    ),
-    icon: <IconSummary />,
-  };
-};
+): Tab => ({
+  label: "Summary",
+  value: detailPageDefault.SUMMARY,
+  component: (
+    <>
+      <SummaryAndDownloadPanel />
+      <MapPanel mapFocusArea={mapFocusArea} onMapMoveEnd={onMapMoveEnd} />
+    </>
+  ),
+  icon: <IconSummary />,
+});
+
+const mapPanelTab = (
+  mapFocusArea: LngLatBounds | undefined,
+  onMapMoveEnd?: (evt: MapEvent) => void
+): Tab => ({
+  label: "Map",
+  value: detailPageDefault.MAP,
+  component: (
+    <MapPanel mapFocusArea={mapFocusArea} onMapMoveEnd={onMapMoveEnd} />
+  ),
+  icon: <IconMap />,
+});
 
 const ContentSection: FC<ContentSectionProps> = ({
   mapFocusArea,
@@ -86,16 +106,27 @@ const ContentSection: FC<ContentSectionProps> = ({
 }) => {
   const { uuid } = useParams();
   const tabNavigation = useTabNavigation();
+  const { isMobile } = useBreakpoint();
 
   const TABS: Tab[] = useMemo(
-    () => [
-      summaryAndDownloadPanelTab(mapFocusArea, onMapMoveEnd),
-      dataAccessPanelTab,
-      citationPanelTab,
-      additionalInfoPanelTab,
-      associatedRecordsPanelTab,
-    ],
-    [mapFocusArea, onMapMoveEnd]
+    () =>
+      isMobile
+        ? [
+            summaryOnlyTab,
+            mapPanelTab(mapFocusArea, onMapMoveEnd),
+            dataAccessPanelTab,
+            citationPanelTab,
+            additionalInfoPanelTab,
+            associatedRecordsPanelTab,
+          ]
+        : [
+            summaryWithMapTab(mapFocusArea, onMapMoveEnd),
+            dataAccessPanelTab,
+            citationPanelTab,
+            additionalInfoPanelTab,
+            associatedRecordsPanelTab,
+          ],
+    [mapFocusArea, onMapMoveEnd, isMobile]
   );
 
   const location = useLocation();
